requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.119
botocore==1.34.119
certifi==2024.6.2
charset-normalizer==3.3.2
frozendict==2.4.4
greenlet==3.0.3
html5lib==1.1
idna==3.7
jmespath==1.0.1
lxml==5.2.2
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==2.2.2
peewee==3.17.5
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
tzdata==2024.1
urllib3==1.26.18
webencodings==0.5.1
yfinance==0.2.40
